ushort indices for vaoMesh

+++ Ottd3DMapEditor.cs
@@ -235,15 +235,15 @@ namespace Ottd3D
 		public void initGrid()
 		{
 			const float z = 0.0f;
-			const int IdxPrimitiveRestart = int.MaxValue;
+			const ushort IdxPrimitiveRestart = ushort.MaxValue;
 
 			Vector3[] positionVboData;
-			int[] indicesVboData;
+			ushort[] indicesVboData;
 			Vector2[] texVboData;
 
 			positionVboData = new Vector3[_gridSize * _gridSize];
 			texVboData = new Vector2[_gridSize * _gridSize];
-			indicesVboData = new int[(_gridSize * 2 + 1) * _gridSize];
+			indicesVboData = new ushort[(_gridSize * 2 + 1) * _gridSize];
 
 			for (int y = 0; y < _gridSize; y++) {
 				for (int x = 0; x < _gridSize; x++) {				
@@ -251,8 +251,8 @@ namespace Ottd3D
 					texVboData [_gridSize * y + x] = new Vector2 ((float)x*0.5f, (float)y*0.5f);
 
 					if (y < _gridSize-1) {
-						indicesVboData [(_gridSize * 2 + 1) * y + x*2] = _gridSize * y + x;
-						indicesVboData [(_gridSize * 2 + 1) * y + x*2 + 1] = _gridSize * (y+1) + x;
+						indicesVboData [(_gridSize * 2 + 1) * y + x*2] =(ushort)(_gridSize * y + x);
+						indicesVboData [(_gridSize * 2 + 1) * y + x*2 + 1] = (ushort)(_gridSize * (y+1) + x);
 					}
 
 					if (x == _gridSize-1) {
